+++ osxphotos/queryoptions.py
@@ -1,12 +1,21 @@
 """ QueryOptions class for PhotosDB.query """
 
+import dataclasses
 import datetime
 from dataclasses import asdict, dataclass
 from typing import Iterable, List, Optional, Tuple
 
 import bitmath
 
-__all__ = ["QueryOptions"]
+from osxphotos.cli.common import load_uuid_from_file
+
+__all__ = ["QueryOptions", "query_options_from_kwargs", "IncompatibleQueryOptions"]
+
+
+class IncompatibleQueryOptions(Exception):
+    """Incompatible query options"""
+
+    pass
 
 
 @dataclass
@@ -182,3 +191,105 @@ class QueryOptions:
 
     def asdict(self):
         return asdict(self)
+
+
+def query_options_from_kwargs(**kwargs) -> QueryOptions:
+    """Validate query options and create a QueryOptions instance"""
+    # sanity check input args
+    nonexclusive = [
+        "added_after",
+        "added_before",
+        "added_in_last",
+        "album",
+        "duplicate",
+        "exif",
+        "external_edit",
+        "folder",
+        "from_date",
+        "from_time",
+        "has_raw",
+        "keyword",
+        "label",
+        "max_size",
+        "min_size",
+        "name",
+        "person",
+        "query_eval",
+        "query_function",
+        "regex",
+        "selected",
+        "to_date",
+        "to_time",
+        "uti",
+        "uuid",
+        "uuid_from_file",
+        "year",
+    ]
+    exclusive = [
+        ("burst", "not_burst"),
+        ("cloudasset", "not_cloudasset"),
+        ("edited", "not_edited"),
+        ("favorite", "not_favorite"),
+        ("has_comment", "no_comment"),
+        ("has_likes", "no_likes"),
+        ("hdr", "not_hdr"),
+        ("hidden", "not_hidden"),
+        ("in_album", "not_in_album"),
+        ("incloud", "not_incloud"),
+        ("is_reference", "not_reference"),
+        ("keyword", "no_keyword"),
+        ("live", "not_live"),
+        ("location", "no_location"),
+        ("missing", "not_missing"),
+        ("only_photos", "only_movies"),
+        ("panorama", "not_panorama"),
+        ("portrait", "not_portrait"),
+        ("screenshot", "not_screenshot"),
+        ("selfie", "not_selfie"),
+        ("shared", "not_shared"),
+        ("slow_mo", "not_slow_mo"),
+        ("time_lapse", "not_time_lapse"),
+    ]
+    # print help if no non-exclusive term or a double exclusive term is given
+    # TODO: add option to validate requiring at least one query arg
+    if any(all([kwargs[b], kwargs[n]]) for b, n in exclusive) or any(
+        # title, description, place, keyword can be specified multiple times
+        # check if any of them are specified along with their no_ counterpart
+        [
+            all([any(kwargs["title"]), kwargs["no_title"]]),
+            all([any(kwargs["description"]), kwargs["no_description"]]),
+            all([any(kwargs["place"]), kwargs["no_place"]]),
+            all([any(kwargs["keyword"]), kwargs["no_keyword"]]),
+        ]
+    ):
+        raise IncompatibleQueryOptions
+
+    # can also be used with --deleted/--not-deleted which are not part of
+    # standard query options
+    try:
+        if kwargs["deleted"] and kwargs["not_deleted"]:
+            raise IncompatibleQueryOptions
+    except KeyError:
+        pass
+
+    # actually have something to query
+    include_photos = True
+    include_movies = True  # default searches for everything
+    if kwargs["only_movies"]:
+        include_photos = False
+    if kwargs["only_photos"]:
+        include_movies = False
+
+    # load UUIDs if necessary and append to any uuids passed with --uuid
+    uuid = None
+    if kwargs["uuid_from_file"]:
+        uuid_list = list(kwargs["uuid"])  # Click option is a tuple
+        uuid_list.extend(load_uuid_from_file(kwargs["uuid_from_file"]))
+        uuid = tuple(uuid_list)
+
+    query_fields = [field.name for field in dataclasses.fields(QueryOptions)]
+    query_dict = {field: kwargs.get(field) for field in query_fields}
+    query_dict["photos"] = include_photos
+    query_dict["movies"] = include_movies
+    query_dict["uuid"] = uuid
+    return QueryOptions(**query_dict)
